DBZ-8964 move postProcessor
